commit 20-04-2022
-Penggabungan REST Controller dan Controller biasa pada Entity Mobil, Penyewa dan Biaya selesai
-Controller biasa menggunakan url ("/") sedangkan REST tidak sehingga lebih mudah di panggil di postman
-CRUD melalui form dan API dapat dijalankan dalam satu Mapping URL

+++ src/main/java/com/rentalmobil/mobil/MobilRepository.java
@@ -3,5 +3,5 @@ package com.rentalmobil.mobil;
 import org.springframework.data.repository.CrudRepository;
 
 public interface MobilRepository extends CrudRepository<Mobil, Integer> {
-    public Long countById(Integer id);
+    public Integer countById(Integer id);
 }
