+++ src/main/java/com/rentalmobil/penyewa/PenyewaRepository.java
@@ -3,5 +3,5 @@ package com.rentalmobil.penyewa;
 import org.springframework.data.repository.CrudRepository;
 
 public interface PenyewaRepository extends CrudRepository<Penyewa, Integer> {
-    public Long countById(Integer id);
+    public Integer countById(Integer id);
 }
